+++ nnpz/tests/python/reference_sample/test_MontecarloDataProvider.py
@@ -104,7 +104,7 @@ def test_appendMc_success(mc_data_files_fixture):
 
     # Given
     expected_data = np.zeros(100, dtype=[
-        ('A', np.float), ('B', np.float), ('C', np.float), ('D', np.float)
+        ('A', np.float32), ('B', np.float32), ('C', np.float32), ('D', np.float32)
     ])
     for c in expected_data.dtype.names:
         expected_data[c] = np.random.rand(*expected_data.shape)
@@ -166,7 +166,7 @@ def test_appendBulk(mc_data_files_fixture):
 
     # Given
     expected_data = np.zeros((20, 100), dtype=[
-        ('A', np.float), ('B', np.float), ('C', np.float), ('D', np.float)
+        ('A', np.float32), ('B', np.float32), ('C', np.float32), ('D', np.float32)
     ])
     for c in expected_data.dtype.names:
         expected_data[c] = np.random.rand(*expected_data.shape)
